sikayetvar_analiz/requirements.txt
flask==3.0.3
scrapy==2.11.2
playwright==1.46.0
google-generativeai==0.7.2
python-dotenv==1.0.1
matplotlib==3.9.2
requests==2.32.3
